Added automatic cut of things chain in case of infinite loop.
Some original DK code tends to break thing chains with invalid effects
removal and creation. This commit adds a function which is able
to cut such chain when infinite loop is detected. While the game
is able to detect most infinite loops, it can't do it with all,
so an invalid chain would sooner or later cause the game to hung.

Cutting the chain may cause some things to be lost from the game,
but it is still better than hung.

Also some unifications in tunnelling wallhug code.